driver/requirements.txt
dataclass_jsonable==0.0.12
flask
RPi.GPIO
